utils: split off pre_filters to separate module
