+++ tests/unit/test_filterout_utils.py
@@ -25,8 +25,7 @@ from __future__ import absolute_import, division, print_function
 from json_merger.conflict import Conflict
 
 from inspire_json_merger.utils import filter_conflicts, \
-    filter_conflicts_by_path, is_to_delete, conflict_to_list, \
-    filter_documents_same_source
+    filter_conflicts_by_path, is_to_delete, conflict_to_list
 
 
 def test_conflict_to_list():
@@ -142,172 +141,3 @@ def test_filter_conflicts():
         'report_numbers'
     ]
     assert len(filter_conflicts(conflicts, fields)) is 4
-
-
-def test_filter_documents_same_source():
-    root = {}
-    head = {
-        'documents': [
-            {
-                'source': 'arXiv',
-                'key': 'file1.pdf',
-                'url': '/files/1234-1234-1234-1234/file1.pdf',
-            },
-            {
-                'source': 'arXiv',
-                'key': 'file2.pdf',
-                'url': '/files/1234-1234-1234-1234/file2.pdf',
-            },
-            {
-                'source': 'publisher',
-                'key': 'file3.pdf',
-                'url': '/files/1234-1234-1234-1234/file3.pdf',
-            },
-        ],
-    }
-    update = {
-        'documents': [
-            {
-                'source': 'arXiv',
-                'key': 'new_file.pdf',
-                'url': '/files/5678-5678-5678-5678/new_file.pdf',
-            },
-        ],
-    }
-    expected_head = {
-        'documents': [
-            {
-                'source': 'publisher',
-                'key': 'file3.pdf',
-                'url': '/files/1234-1234-1234-1234/file3.pdf',
-            },
-        ],
-    }
-
-    result = filter_documents_same_source(root, head, update)
-    expected = root, expected_head, update
-
-    assert result == expected
-
-
-def test_filter_documents_same_source_multiple_sources_in_update():
-    root = {}
-    head = {
-        'documents': [
-            {
-                'source': 'arXiv',
-                'key': 'old_file.pdf',
-                'url': '/files/5678-5678-5678-5678/old_file.pdf',
-            },
-        ],
-    }
-    update = {
-        'documents': [
-            {
-                'source': 'arXiv',
-                'key': 'file1.pdf',
-                'url': '/files/1234-1234-1234-1234/file1.pdf',
-            },
-            {
-                'source': 'arXiv',
-                'key': 'file2.pdf',
-                'url': '/files/1234-1234-1234-1234/file2.pdf',
-            },
-            {
-                'source': 'publisher',
-                'key': 'file3.pdf',
-                'url': '/files/1234-1234-1234-1234/file3.pdf',
-            },
-        ],
-    }
-
-    result = filter_documents_same_source(root, head, update)
-    expected = root, head, update
-
-    assert result == expected
-
-
-def test_filter_documents_remove_head_source():
-    root = {}
-    head = {
-        'documents': [
-            {
-                'source': 'publisher',
-                'key': 'file3.pdf',
-                'url': '/files/1234-1234-1234-1234/file3.pdf',
-            },
-            {
-                'source': 'arXiv',
-                'key': 'old_file.pdf',
-                'url': '/files/5678-5678-5678-5678/old_file.pdf',
-            },
-        ],
-    }
-    update = {
-        'documents': [
-            {
-                'source': 'publisher',
-                'key': 'file3.pdf',
-                'url': '/files/1234-1234-1234-1234/file3.pdf',
-            },
-
-        ],
-    }
-    expected_head = {
-        'documents': [
-            {
-                'source': 'arXiv',
-                'key': 'old_file.pdf',
-                'url': '/files/5678-5678-5678-5678/old_file.pdf',
-            },
-        ],
-    }
-
-    result = filter_documents_same_source(root, head, update)
-    expected = root, expected_head, update
-
-    assert result == expected
-
-
-def test_filter_documents_same_source_is_case_insensitive_on_source():
-    root = {}
-    head = {
-        'documents': [
-            {
-                'source': 'arXiv',
-                'key': 'file1.pdf',
-                'url': '/files/1234-1234-1234-1234/file1.pdf',
-            },
-            {
-                'source': 'arXiv',
-                'key': 'file2.pdf',
-                'url': '/files/1234-1234-1234-1234/file2.pdf',
-            },
-            {
-                'key': 'file3.pdf',
-                'url': '/files/1234-1234-1234-1234/file3.pdf',
-            },
-        ],
-    }
-    update = {
-        'documents': [
-            {
-                'source': 'arxiv',
-                'key': 'new_file.pdf',
-                'url': '/files/5678-5678-5678-5678/new_file.pdf',
-            },
-        ],
-    }
-    expected_head = {
-        'documents': [
-            {
-                'key': 'file3.pdf',
-                'url': '/files/1234-1234-1234-1234/file3.pdf',
-            },
-        ],
-    }
-
-    result = filter_documents_same_source(root, head, update)
-    expected = root, expected_head, update
-
-    assert result == expected
